improve ergonomics of specifying plot legend
