+++ crates/re_types/definitions/rerun/blueprint/archetypes/plot_legend.fbs
@@ -13,7 +13,8 @@ namespace rerun.blueprint.archetypes;
 /// Configuration for the legend of a plot.
 table PlotLegend (
     "attr.rerun.scope": "blueprint",
-    "attr.rust.derive": "Default"
+    "attr.rust.derive": "Default",
+    "attr.python.aliases": "blueprint_components.Corner2D"
 ) {
     // --- Optional ---
