+++ rerun_py/rerun_sdk/rerun/blueprint/archetypes/plot_legend.py
@@ -5,48 +5,20 @@
 
 from __future__ import annotations
 
-from typing import Any
-
 from attrs import define, field
 
 from ..._baseclasses import Archetype
 from ...blueprint import components as blueprint_components
-from ...error_utils import catch_and_log_exceptions
+from .plot_legend_ext import PlotLegendExt
 
 __all__ = ["PlotLegend"]
 
 
 @define(str=False, repr=False, init=False)
-class PlotLegend(Archetype):
+class PlotLegend(PlotLegendExt, Archetype):
     """**Archetype**: Configuration for the legend of a plot."""
 
-    def __init__(
-        self: Any,
-        *,
-        corner: blueprint_components.Corner2DLike | None = None,
-        visible: blueprint_components.VisibleLike | None = None,
-    ):
-        """
-        Create a new instance of the PlotLegend archetype.
-
-        Parameters
-        ----------
-        corner:
-            To what corner the legend is aligned.
-
-            Defaults to the right bottom corner.
-        visible:
-            Whether the legend is shown at all.
-
-            True by default.
-
-        """
-
-        # You can define your own __init__ function as a member of PlotLegendExt in plot_legend_ext.py
-        with catch_and_log_exceptions(context=self.__class__.__name__):
-            self.__attrs_init__(corner=corner, visible=visible)
-            return
-        self.__attrs_clear__()
+    # __init__ can be found in plot_legend_ext.py
 
     def __attrs_clear__(self) -> None:
         """Convenience method for calling `__attrs_init__` with all `None`s."""
